+++ lab1/src/StandardMultiplication.java
@@ -1,5 +1,5 @@
 class StandardMultiplication {
-    public static int operationCount = 0;
+    public static long operationCount = 0;
 
     public static double[][] multiply(double[][] A, double[][] B) {
         int n = A.length;
